Create pie.py
One more task for the brain work out!
There are 2 teams (a and b) which are participate in a competition. The winner will get a pie as a reward. You need to create a program which counts for how many pieces you need to divide a pie. Each member of the winning team should get the the same number of pieces.